refactor BulkWriter's logic in UpdateBuilder into BulkCommitBatch
